requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.6
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cmake==4.3.0
colorama==0.4.6
cryptography==46.0.1
cvzone==1.6.1
deprecation==2.1.0
dlib @ file:///C:/Users/HP/Desktop/Attendence/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.8.0
email-validator==2.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-SocketIO==5.6.1
frozenlist==1.7.0
fsspec==2026.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.6.4
numpy==1.26.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
postgrest==2.28.3
propcache==0.3.2
py-vapid==1.9.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.10.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
pywebpush==2.0.3
realtime==2.28.3
requests==2.32.5
rich==14.3.3
simple-websocket==1.1.0
six==1.17.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
webpush==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
yarl==1.23.0
zstandard==0.25.0
